// src/types.ts
export enum TravelMode {
  DRIVING = 0,
  CYCLING = 1,
  WALKING = 2,
  TWO_WHEELER = 3,
}
// ...
export type LatLng = {
  latitude: number;
  longitude: number;
};

export type CameraPosition = {
  target: LatLng;
  zoom?: number;
  bearing?: number;
  tilt?: number;
};

export type Waypoint = {
  position: LatLng;
  title?: string;
  placeId?: string;
  metadata?: Record<string, string>;
};